Source: https://github.com/msupply-foundation/open-msupply/commit/b75e9a677c61271ddacbe6d851c5a15e77b47303

+++ server/graphql/demographic/src/mutations/update_projection.rs
@@ -49,11 +49,11 @@ pub fn update_demographic_projection(
 pub struct UpdateDemographicProjectionInput {
     pub id: String,
     pub base_year: Option<i32>,
-    pub year_1: Option<i32>,
-    pub year_2: Option<i32>,
-    pub year_3: Option<i32>,
-    pub year_4: Option<i32>,
-    pub year_5: Option<i32>,
+    pub year_1: Option<f64>,
+    pub year_2: Option<f64>,
+    pub year_3: Option<f64>,
+    pub year_4: Option<f64>,
+    pub year_5: Option<f64>,
 }
 
 impl From<UpdateDemographicProjectionInput> for UpdateDemographicProjection {
